+++ python-blutils/RELEASENOTES.md
@@ -1,5 +1,20 @@
 # Release Notes
 
+### 1.3.0 (2022-12-22)
+* Calling with no parameters now outputs status
+* Support --help, --version; deprecating --list
+* Support executing from different cwd.
+
+### 1.2.0 (2022-12-27)
+* Added --list to list configured deployables
+* Removed utils1 file (DEPLOY.py now single-file)
+
+### 1.1.0 (2022-12-22)
+* 1 log line per deploy (not multiple); made repoDir required in config.
+* Merged appName & envName into single identifier, appEnvName.
+
 ### 1.0.0 (2022-12-02)
 * initial state, formerly utils.py
+
+